fix post place
